+++ src/test/java/net/rentalhost/plugins/php/hammer/inspections/codeStyle/ParameterDefaultsNullInspectionTestCase.kt
@@ -1,0 +1,7 @@
+package net.rentalhost.plugins.php.hammer.inspections.codeStyle
+
+import net.rentalhost.plugins.services.TestCase
+
+class ParameterDefaultsNullInspectionTestCase: TestCase() {
+    fun testAll(): Unit = testInspection(ParameterDefaultsNullInspection::class.java)
+}